- Added solution for when walking_dependencies in inki environment   after loading a file, and we have pointers with a type that is a   type root.
The walk dependencies function would be called to only handle the
root type, and not the subclasses.

Added flag to function that says if it came from a higher up type,
and if so it will not call down.

If not set it will try to call down to the real type.